+++ package.json
@@ -13,7 +13,8 @@
     "multer": "^1.4.5-lts.1",
     "next": "^14.2.4",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-select": "^5.8.0"
   },
   "devDependencies": {
     "@types/node": "20.14.9",
