--- src/app/components/ProfileBusiness.tsx
@@ -1,13 +1,0 @@
-
-
-
-export default function ProfileBusiness() {
-    return (
-      <div>
-        <h1>Profile Business</h1>
-        <p>This is not arbitrary javascript, you freaking fool
-            </p>
-      </div>
-    );
-  }
-
