--- src/app/components/ProfilesUser.tsx
@@ -1,10 +1,0 @@
-
-
-export default function ProfileUser() {
-    return (
-      <div>
-        <h1>Profile User </h1>
-      </div>
-    );
-  }
-
