+++ src/app/index.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Cabin:ital,wght@0,400..700;1,400..700&display=swap');
+
 body {
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
       'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
@@ -44,6 +46,7 @@ body {
     flex-direction: column;
   }
 
+  
 
   /* ================
    TOP AND NAVBAR ======================================================
@@ -318,7 +321,7 @@ body {
   }
 
     /* SPECIAL FONTS */
-    @import url('https://fonts.googleapis.com/css2?family=Cabin:ital,wght@0,400..700;1,400..700&display=swap');
+
     .cabinfont {
     font-family: "Cabin", sans-serif;
     font-optical-sizing: auto;
@@ -479,11 +482,29 @@ body {
     .buttonaf {
     display: flex;
     flex-direction: column;
-
     align-items: center;
     padding: 10px;
     justify-content: space-evenly;
     margin: 20px;
-
     max-width: 250px;
+  }
+
+  .active {
+    background-image: linear-gradient( #9700ff, aqua, blue, #9700ff);
+
+  }
+  .show {
+    background-image: linear-gradient( #9700ff, aqua, blue, #9700ff);
+  }
+
+  .themebutton {
+    background-image: linear-gradient( #9700ff, aqua, blue, #9700ff);
+  }
+
+  .dark-theme {
+      background-color: black;
+  }
+
+  .theme_picker {
+    height: 300px;
   }
